+++ app/(quiz)/welcome.jsx
@@ -17,17 +17,17 @@ export default function Welcome() {
         >
             <View className="flex-1 bg-black/40">
                 <SafeAreaView className="flex-1 justify-center align-center items-center px-10 ">
-                    <View className="w-full flex-column items-center mb-16 mt-32 gap-y-1">
-                        <Text className="text-5xl font-bold JakartaBold text-white text-center mb-0">
-                            Rich Spirit
+                    <View className="w-full flex-column items-center mb-2 mt-32 gap-y-1">
+                        <Text className="text-4xl font-bold JakartaBold text-white text-center mb-0">
+                            Discover your tribe
                         </Text>
                         <Text className="text-white text-md font-medium JakartaMedium text-center mb-8">
-                            Discover your tribe. Unlock the spirit animal that guides you
+                            Unlock the animal spirit that guides your financial decisions
                         </Text>
                     </View>
 
                     <TouchableOpacity
-                        className="bg-yellow-700 px-8 py-4 rounded-full"
+                        className="bg-[#fd7e14] px-8 py-4 rounded-full"
                         onPress={handleFindTribe}
                     >
                         <Text className="text-white text-lg font-medium JakartaMedium">